+++ src/test/java/picard/util/AlleleSubsettingUtilsTest.java
@@ -1,16 +1,16 @@
 package picard.util;
 
 import htsjdk.variant.variantcontext.*;
-import htsjdk.variant.vcf.VCFConstants;
 import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
-import picard.fingerprint.FingerprintChecker;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+
+import static picard.util.AlleleSubsettingUtils.NON_REF_ALLELE;
 
 public class AlleleSubsettingUtilsTest {
 
@@ -134,9 +134,9 @@ public class AlleleSubsettingUtilsTest {
                 {1, Arrays.asList(Aref, C, G), new double[]{0, 2, 5}, Arrays.asList(Aref, G)}, //second is best
                 {1, Arrays.asList(Aref, C, G), new double[]{0, 1, 1}, Arrays.asList(Aref, C)}, //tie chooses first
                 {1, Arrays.asList(Aref, C, G), new double[]{5, 1, 1}, Arrays.asList(Aref, C)}, //ref score is ignored chooses first
-                {2, Arrays.asList(Aref, C, FingerprintChecker.NON_REF_ALLELE, G), new double[]{0, 5, 0, 2}, Arrays.asList(Aref, C, FingerprintChecker.NON_REF_ALLELE, G)}, //keep NON_REF in order
-                {1, Arrays.asList(Aref, C, FingerprintChecker.NON_REF_ALLELE, G), new double[]{0, 5, 0, 2}, Arrays.asList(Aref, C, FingerprintChecker.NON_REF_ALLELE)}, //keep NON_REF in order when trimming
-                {1, Arrays.asList(Aref, C, FingerprintChecker.NON_REF_ALLELE, G), new double[]{0, 5, 0, 2}, Arrays.asList(Aref, C, FingerprintChecker.NON_REF_ALLELE)}, //keep NON_REF in order when trimming
+                {2, Arrays.asList(Aref, C, NON_REF_ALLELE, G), new double[]{0, 5, 0, 2}, Arrays.asList(Aref, C, NON_REF_ALLELE, G)}, //keep NON_REF in order
+                {1, Arrays.asList(Aref, C, NON_REF_ALLELE, G), new double[]{0, 5, 0, 2}, Arrays.asList(Aref, C, NON_REF_ALLELE)}, //keep NON_REF in order when trimming
+                {1, Arrays.asList(Aref, C, NON_REF_ALLELE, G), new double[]{0, 5, 0, 2}, Arrays.asList(Aref, C, NON_REF_ALLELE)}, //keep NON_REF in order when trimming
         };
     }
 }
